microservicio/infraestructura/src/test/java/com/ceiba/precio/controlador/ConsultaControladorPrecioTest.java
package com.ceiba.precio.controlador;

import com.ceiba.ApplicationMock;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.http.MediaType;
import org.springframework.test.annotation.DirtiesContext;
import org.springframework.test.context.ContextConfiguration;
import org.springframework.test.context.junit.jupiter.SpringExtension;
import org.springframework.test.web.servlet.MockMvc;

import static org.hamcrest.Matchers.hasSize;
import static org.hamcrest.core.Is.is;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

@ExtendWith(SpringExtension.class)
@WebMvcTest(ConsultaControladorPrecio.class)
@ContextConfiguration(classes = ApplicationMock.class)
@DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_EACH_TEST_METHOD)
public class ConsultaControladorPrecioTest {

    @Autowired
    private MockMvc mockMvc;

    @Test
    @DisplayName("Deberia listar los precios de las habitaciones")
    void deberiaListarLosPreciosDeLasHabitaciones() throws Exception {
        // arrange
        // act - assert
        mockMvc.perform(get("/precios")
                        .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isOk())
                .andExpect(jsonPath("$", hasSize(3)))
                .andExpect(jsonPath("$[0].precioSemana", is(180000)))
                .andExpect(jsonPath("$[0].id", is(1)));
    }

    @Test
    @DisplayName("Deberia listar el precio de una habitación en especifico")
    void deberiaListarPreciosPorIdTipoHabitacion() throws Exception {
        // arrange
        Long idTipoHabitacion = 1L;
        // act - assert
        mockMvc.perform(get("/precios/{idTipoHabitacion}", idTipoHabitacion)
                        .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isOk())
                .andExpect(jsonPath("$", hasSize(1)))
                .andExpect(jsonPath("$[0].precioSemana", is(180000)))
                .andExpect(jsonPath("$[0].id", is(1)));

    }
}
